Move extractDependencies under resolver
