+++ src/mirage/resolvers/root-query.ts
@@ -10,7 +10,7 @@ import {
   RootQueryResolverMatch,
   AutoResolverErrorMeta,
 } from './helpers';
-import { extractDependencies } from '../../resolver-map/extract-dependencies';
+import { extractDependencies } from '../../resolver/extract-dependencies';
 import { AutoResolverError } from './auto-resolver-error';
 
 function findMatchingModelsForType({
