+++ src/resolver/extract-dependencies.ts
