+++ test/integration/test-helpers/mirage-static-resolvers/Query/all-persons.ts
@@ -1,5 +1,5 @@
 import { ModelInstance, Server } from 'miragejs';
-import { extractDependencies } from '../../../../../src/resolver-map/extract-dependencies';
+import { extractDependencies } from '../../../../../src/resolver/extract-dependencies';
 
 export default function (
   _parent: unknown,
